pyproject.toml
 [project]
 name = "sisy-designs"
 version = "0.1.0"
 description = "Tools for reproducing SiSy design screenshots."
 readme = "docs/v0.md"
 requires-python = ">=3.11"
 dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "langchain>=0.1.0",
  "langchain-openai>=0.0.5",
  "opik>=1.0.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "requests>=2.31.0",
  "python-multipart>=0.0.22",
  "mcp>=1.26.0",
  "minimax-coding-plan-mcp>=0.0.3",
  "playwright>=1.41",
 ]
 
